Get most tests passing with self-compiled compiler. This was a really big step!
There were issues with genTernary() in gen.c that was freeing
registers that it wasn't supposed to be freeing, and this was
leading to bugs in lines like:

d += a > b ? c : d;

Also, there was another call to freeall_registers() where it wasn't
supposed to be, and cgcompare_and_jump. Now it only frees the
appropriate registers.